Templated floats with gradcheck support for double precision

+++ test/utils.py
@@ -11,7 +11,8 @@ def sample_1d_inputs(device, *, requires_grad=False):
 
 def sample_2d_inputs(device, *, requires_grad=False):
     def make_tensor(*size):
-        return torch.randn(size, device=device, requires_grad=requires_grad)
+        return torch.randn(size, dtype=torch.float, device=device, requires_grad=requires_grad)
+            
 
     return [
         [make_tensor(32,32), make_tensor(32,32)],
